Work around Clang UBSAN conversion bug on armv8
Summary:
When compiling with `-fsanitize=undefined` on armv8, Clang wrongly
performs a signed byte load (`ldursb`) instead of an unsigned load
(`ldurb`). Because the value gets sign extended, it seems like we are
reading past the end of the buffer. This means that the following
program is sufficient to trigger a UBSAN error:

```
#include <cstdint>

int main(){
  uint8_t opCode = 128;
  void *opCodeStrings[200];
  opCodeStrings[(int)opCode];
  return 0;
}
```

See https://godbolt.org/z/f5cK6ea7T for the difference in generated
code between the old version and the new version.

Filed an issue on LLVM to follow up:
https://github.com/llvm/llvm-project/issues/59630

Reviewed By: tmikov

Differential Revision: D42184531

fbshipit-source-id: f375e8002fd3e037a2616c837ef0d598453ae07c