Improve local eval() semantics a bit
Summary:
Local eval() is supposed to be a normal call if the identifier "eval"
resolves to something other than the builtin eval, and is supposed to be
a direct eval if it does resolve. In the former cases it must be able to
support any number of arguments.

Simplify the checking for local eval: at compile time it is sufficient
to know that the identifier is called "eval". We will its actual value
at runtime. Note that this means that this will *correctly* be treated
as a direct eval:

   let eval = globalThis.eval;
   let p = 10;
   eval("print(p)");

Move the checking of whether the callee is actually `eval` or not
outside of the DirectEval instruction. This needs to happen so we can
call a non-eval function just like a normal function. The check relies
on the recently added "normal builtin" `eval`.

While we are here, make sure eval() inherits the surrounding strictness,
since that was already added to Hermes.

Reviewed By: neildhar

Differential Revision: D42225222

fbshipit-source-id: 00d45684db0b9ead1fe541d1b6c090c3ddb84a50