SemanticResolver: don't resolve the name of class members
Summary:
We were visiting the identifier of class members (private properties,
properties, methods) and attempting to resolve it as a variable in the
scope surrounding the class. Simply skip the identifier (note that in
the future we may need to resolve private property names for "Legacy JS
class support").

The property initialization value, if present, may have to be resolved
differently, as if inside a new synthetic method. The value technically
has access to `this`, but nothing else. But we technically don't resolve
`this`.

For now postpone this decision and simply error out if there is an init
value, so at least we don't potentially resolve it incorrectly.

Reviewed By: avp

Differential Revision: D42492192

fbshipit-source-id: 925c20bdc6f90482a5ee860cfb42c984d8b76e91