shermes: enable running and dumping of the new resolver
Summary:
Before we enable the new resolver, we need the ability to run it and
dump its output for testing, without actually using it.

This requires temporarily adding some "new" attributes to the AST,
separate from the ones used by the "old" resolver. They will be removed
once the new resolver is activated.

Reviewed By: avp

Differential Revision: D42527129

fbshipit-source-id: f8d2faf6781258227f5e66cff7cd5a6d3e33c8a6