Add a utility function for deleting unused functions and variables
Summary:
Add a new function that deletes all unused functions and variables, as
well as a helper that deletes just unused variables. The distinction is
that we always want to delete unused variables when we delete unused
functions, but there may be instances were we just want to do the
latter (e.g. in StackPromotion).

The code for deleting unused functions is copied from DCE, and the
original code in DCE is replaced with a call to the new utility.

Reviewed By: tmikov

Differential Revision: D42831466

fbshipit-source-id: 8ae83cb9a7a80ac4360a68cfeeed7c57440982a8