Add a command line option for creating an individual unit
Summary:
Add an `-exported-unit` command line option that allows a user to
specify that they want to export the produced `SHUnit` with a given
name, rather than trying to produce an executable.

Whether to emit a main function is based on whether this flag is
specified, as opposed to the output type. Separating the two is useful
for the C output type in particular, since we may want it with or
without the main function included.

Reviewed By: tmikov

Differential Revision: D42616179

fbshipit-source-id: 97cd31a0b66aeeec9a38bda89177e6b33fd7c69f