Compile InternalBytecode to native code
Summary:
Compile the promise polyfill to native SH code.

This works by invoking the `shermes` compiler from at build time in
order to compile the InternalBytecode JS source to a single C file.
That C file is then incorporated into the build, and linked into
libhermesvm. By using a C file instead of producing an object file
directly, we don't have to deal with matching the compiler and
target platform of the CMake build.

Reviewed By: tmikov

Differential Revision: D42616176

fbshipit-source-id: c6601352bdbfc140f440b50e998789e496e50a7f