ADT/ScopedHashTable: add eraseFromCurrentScope()
Summary:
`eraseFromCurrentScope()` does what it name suggests. It can be slow,
since it needs to scan the singly linked list of the current scope to
find the node before the one we are erasing.

However, in the next diff it is used only in a context where it will be
fast. First, the scope in question will only contain parameters, which
are usually only a few. Second, we will be calling it to delete the last
inserted parameter, which is always first in the list.

Reviewed By: avp

Differential Revision: D42952094

fbshipit-source-id: 7c67303d0414dc09808c06da9e412eb7e51e8fb0