Pass Instruction to isDirectCallee
Summary:
`isDirectCallee` is used for checking that the callee is the operand
which contains the potential callee, as well as whether it's captured.

Now that we've added the `target` operand which creates another use
of the `Function`, we will stop calling `isDirectCallee` with a
`Function *` callee, so codify that in the type system by using
`Instruction *` instead.

Reviewed By: neildhar

Differential Revision: D40965235

fbshipit-source-id: 14c8c2b387ca0dcb2ce7404bdbbbddb6d6ebeb76