Add function and environment operands to CallInst
Summary:
Add extra operands to CallInst to more clearly track which functions
are being called with which environments. This will be even more useful
as we implement static typing, inlining, optimizations, etc.

The `target` and `env` fields are either populated or `EmptySentinel`,
to indicate they haven't been set yet.

The fields will be populated with an IR optimization pass prior to
passes such as inlining, and will be transferred during lowering.

This diff doesn't even populate the fields, it just always leaves
them empty.

Reviewed By: tmikov

Differential Revision: D40776765

fbshipit-source-id: 1a342948fe726775b54790011612a905c8d3fa5a