runtime: add property load and store by index APIs
Summary:
The new APIs allow faster (and completely unchecked) property loads and
stores by slot index.

First define a few lower level calls:
- `_sh_prload_direct()`, `_sh_prload_indirect()` load from a direct or
  indirect property slot
- `_sh_prstore_direct()`, `_sh_prstore_indirect()` store to a direct or
  indirect property slot.
- `_sh_prstore_direct_np()` store to a direct property slot when it is
  known that the both the new stored value and the old value are not
  pointers. This removes the need for a GC barrier.

Note that there is no `_sh_prstore_indirect_np()`, because there is no
such an underlying API. We could create it, but the feeling is that, at
least for now, writing to indirect properties is already expensive and
relatively rare, so it is not worth worrying about GC barriers. Of
course these tradeoffs will be reevaluated in the future.

Then define a few higher level APIs, as inline functions, automatically
switching between the above APIs, based on the assumption that the
property index is a compile time constant:
- `_sh_prload()`: load from any property slot.
- `_sh_prstore()`: store to any property slot.
- `_sh_prstore_np()`: store to any property slot, when old and new value
  are non-pointers.

Reviewed By: neildhar

Differential Revision: D42492182

fbshipit-source-id: fc242e1d9582e95d7c77bc6ae45ad613c7a14d8b