JSObject: remove PropStorage::Inline setting and split unsafe funcs
Summary:
The PropStorage::Inline setting in a few JSObject methods was confusing
and unnecessary, since it wasn't actually used. Remove it.

Split the unsafe property methods into separate methods for direct and
indirect properties, while leaving the previous method to call either
one.

Add a template parameter for GC write barrier needed to the direct
property store method, to speed up non-pointer updates (when we know
that both the old and the new value are not pointers).

Reviewed By: neildhar

Differential Revision: D42492187

fbshipit-source-id: ab78c539f8e4e0104d1a074ed757d0ddb14a3630