Remove HERMESVM_DEBUG_MODE
Summary:
HERMESVM_DEBUG_MODE was meant to encode the knowledge of whether
libhermesvm was compiled with or without asserts. It was intended to be
similar to other configuration defines.

The problem is that nothing is checking it. Both the library and the
generated client code rely on NDEBUG and assert(). To make it work we
would have to change all references to NDEBUG to HERMESVM_DEBUG_MODE,
and would need to define a custom hermes_assert() depending on the
latter.

But why go to that trouble when NDEBUG and assert() are already there
and are well understood?

So, instead, change libhermesvm-config.h to simply define a different
model name depending on the current setting of NDEBUG. If the library
and the client are compiled with different NDEBUG, they will get a
different _SH_MODEL and a link error.

Additionally, set the default value of --enable-asserts in shermes based
on NDEBUG, so it matches the default shermes library.

IMPORTANT: when you start using this commit, make sure to delete
SHERMES_CC_INCLUDE_PATH from CMakeCache.txt.

Reviewed By: neildhar, avp

Differential Revision: D43338375

fbshipit-source-id: ac1264fca56ba8d893ad1d1464e0ae43c3ffaf8a