New IR instructions PrLoad and PrStore
Summary:
`prload (:type) object, propIndex, propName` loads a property with the
specified index, name and type from an object. The name isn't actually
used, it is added for readability, and may be used in the future for
debugging. The property type is stored as an additional attribute and
propagated during type inference into the instruction type.

`prstore value, object, propIndex, propName, nonPointer` stores a
property in the obvious way. Again, the name isn't used. `nonPointer` is
set to true if it is known at compile time that both the old value of
the property and the new one are not pointers. It would be true for all
statically typed non-pointer properties. The intent is to avoid a GC
write barrier.

Reviewed By: avp

Differential Revision: D42492199

fbshipit-source-id: 05d351449acb3130c6613e6cb5616be39108f67b