Delete PointerBase
Summary:
Make `SHRuntime` a base class of `Runtime`, and get rid of
`PointerBase`, since the former now provides its fields.

This has some drawbacks, since for now, `PointerBase` has been turned
into an alias for `SHRuntime`, in order to allow us to continue to use
the name `PointerBase` in VM code, and easily convert `Runtime` to it.

There were a few alternatives I considered, that I'm also open to
getting feedback on:
1. Make a separate struct `SHPointerBase`, that just stores the fields
needed to do pointer conversion. We would probably need to introduce an
implicit conversion operator for `Runtime` to avoid changing a lot of
existing code.
2. Make `SHRuntime` the base class of `PointerBase`, and then make
`PointerBase` the base class of `Runtime`. This seems like a special
hack just for `PointerBase`.

I figured this approach is simpler and we can always switch later,
given that the API is fairly well defined.

Reviewed By: tmikov

Differential Revision: D43072600

fbshipit-source-id: 3a1e5dad6f647875db752ad58cb3a19ea79edc49