Analyze call graph in new FunctionAnalysis pass
Summary:
Add a `FunctionAnalysis` pass which will allow us to register info
about calls and functions into the IR.

We effectively generate a call graph and encode it in the IR this way.
The created information is really only used by the inliner in this diff,
but it can be expanded to give more information in the future.

This diff makes the analysis pass and populates the target/env fields.

Reviewed By: neildhar

Differential Revision: D40962627

fbshipit-source-id: 3ec44ba24f4369603b35246d265afbb09d4edff7