Flow: change array and function types to structural
Summary:
Two array types are the same if they have the same element type.
Similarly, two function types are the same if they have the same
attributes, parameters and return type. This is unlike classes, which
are the same only if they are the same class.

One implication is for union equality. In the example below, U1 and U2
are the same union:

```
type A1 = number[];
type A2 = number[];
type U1 = string | A1;
type U2 = A2 | A1 | string;
```

As can be seen in the example, unions must be able to coalesce
equivalent arrays and functions, as well as sort them predictably.

Since unions are sorted in FlowContext, it now needs a notion of
ordering structural types. Introduce that and pair it with hashing,
which will be used only for dumping the state for now.

Consequently, remove ArrayType and FunctionType from the TypeWithId
hierarchy. Introduce `compare()` and `hash()` pseudo-virtual methods.

It may be possible to achieve this result by uniquing structural types,
but the code is not ready for such a major refactor. Besides, even
equivalent types will probably still need to carry information like
declaration location.

Modify the FlowTypesDumper to use the new hashing functionality and to
avoid printing equivalent types.

Reviewed By: avp

Differential Revision: D43561748

fbshipit-source-id: bcb3818b250d3061db1d9c30b30ffa5b9e39870e