Widgets: Write specialized JS Array library functions
Summary:
Add versions of JSLib array functions which can be inlined and can
benefit from static typing.

Eventually these functions will be able to be generated with generics,
but for now we just write specializations for each type we care about,
and mark them `"inline"` so we can inline the callbacks and speed
things up significantly.

Reviewed By: neildhar

Differential Revision: D43703197

fbshipit-source-id: 9acf898bd59b34688a9cd068c4721b1eda61c276