Preserve empty-type correctness in the IR
Summary:
Until now, `:any` included `:empty`, which was incorrect, since `:empty`
is not a valid input for most instructions. Rename the types, so that
the new `:any` does not include `:empty` and the "old" `:any` is printed
as `:any|empty`.

Update the code to generate the correct types. `:empty` should now only
be present after a load from a TDZ variable, and should be short-lived.

An operation that cannot handle `:empty` as an operand, should never see
it in its operand types.

It remains to be seen whether TDZDedup has to be changed to maintain
empty-correctness, but that will be in another diff.

Reviewed By: avp

Differential Revision: D43970370

fbshipit-source-id: e8fe87687a861d5e4fb19de53ecd21b085dd67ba