Transition to a SH-specific register allocator
Summary:
For now this is a mechanical change. SHRegisterAllocator is a copy if
HVMRegisterAllocator, in a different directory and namespace. A few
changes in previous diffs were needed to make this possible.

This unlocks meaningful SH-specific code generation in followup diffs,
like introducing different register classes for locals and
register stack.

Reviewed By: avp

Differential Revision: D44008274

fbshipit-source-id: e731d87412a08f2e2bb81702ec4bed38dc97c175