Run executables in-process
Summary:
When the user specifies `-exec`, we currently produce a standalone
executable and then spawn a new process for it. Instead, we can emit a
shared library and then `dlopen` it. This is slightly faster, but more
importantly, makes it easy to debug generated programs, since we can
just use `shermes -exec`.

For now, it uses the exact same C code (including the main function),
and directly invokes the main function.

Some potential further improvements:
1. Expose the `SharedObj` output type to the CLI.
2. Make `sh_init` take `const argv *const *` so it is clearer that it
is const. I skipped that here because `char **` is technically only
compatible with `const char *const *` in C++.
3. Remove the need for `keep-temp` by keeping the source around until
after execution is done.
4. Always build with debug info for `-exec`.

Reviewed By: tmikov

Differential Revision: D44179271

fbshipit-source-id: bd47b46fdf772ce424e2deae006e96602a0db32a