Eliminate UnionNarrowTrusted during lowering
Summary:
The new pass is called `LowerNanBoxedUnionNarrowTrusted`, since its
behavior depends on the value encoding, specifically that all unions and
values have the same Nan-boxed representation.

With a Nan-boxed encoding, UnionNarrowTrusted doesn't do any actual
work, it is only a type assertion. So the new pass deletes it and
narrows the input type to match its result type.

We also need to deal with the theoretical possibility that multiple
UnionNarrowTrusted instructions with different result types are using
the same input value. Since they must all be correct, they must
represent different subsets of the input value, where the input's "true"
type is an intersection of all of them. So, that is what we do.

Finally, we need to contend with the possibility that this is dead code,
which is indicated by the empty intersection of the input and the output
type. In that case we emit `LIRDeadValue` with the appropriate type.

This fixes the last test that was previously disabled by this stack.

Reviewed By: neildhar

Differential Revision: D44406198

fbshipit-source-id: c043803c2fd639279e002e004eeb35060fa9fc03