Fix asserts on GCHermesValueBase::uninitialized_copy
Summary:
These asserts would fire even on correct code, for instance if we're
copying elements from the start of a range right past the end of that
range. e.g. `uninitialized_copy(a, b, b, gc)`.

The names here are a little misleading, since `last` suggests it is a
pointer to the last element, when in fact it is a pointer to the end of
the range, which is likely why the assert was incorrect. However, these
names are used on multiple functions, so just fix the assert for now.

Reviewed By: avp

Differential Revision: D44260263

fbshipit-source-id: a041b969232b8a0fe0c8351fffce600e538b10fb