Delete allReadOnly and allNonConfigurable flags on HiddenClass
Summary:
Original Author: neildhar@meta.com
Original Git: d946927b5988de3bfd4e61e26b186ece8d147384

We currently do not update the class flags when properties are added or
updated, which means that these flags can go out-of-date. For instance,
if a configurable property is added to a class with
`allNonConfigurable` already set, the newly created child class (or the
current class if it is in dictionary mode) will still have that flag
set.

This is tricky to reproduce, since we only cache the value when the
object owning the class has been made non-extensible, so extensions
must be prevented first, and then a property must be added to the
hidden class. This can be observed either by having two objects with a
common hidden class, or by transitioning an indexed property to a named
one.

Discussed offline with tmikov, and we don't believe this optimisation is
likely to make a difference in practice. So get rid of it and simplify the
code. We can add it back later if it seems useful.

Original Reviewed By: tmikov

Original Revision: D41983713

Reviewed By: tmikov

Differential Revision: D44771003

fbshipit-source-id: 26e1619fc0ad85231da5e1a450febe1517297570