Use copyDataProperties slow path for objects with accessors
Summary:
Original Author: fbmal7@meta.com
Original Git: 25674a326addf3bfd6b411de119fb95f17b42082

Fix a bug present in `hermesBuiltinCopyDataProperties`.
`hermesBuiltinCopyDataProperties` is meant to copy an object. It does
this by iterating over all the properties in the source object, and
then writing the key-value pair into a new object which is returned.The
iteration is done using `JSObject::forEachOwnPropertyWhile`, providing
a callback to process each property. It will take the property id, look
it up in the source object using `getNamedPropertyValue_RJS`, then take
that resulting value and write it into the target. However, that API
explicitly states the following:

> Obviously the callbacks shouldn't be  doing naughty things like modifying
the property map or creating new hidden classes (even implicitly)

But, `getNamedPropertyValue_RJS` can
perform arbitrary JS execution if it is a getter, including
adding/deleting new properties, which would modify the object's
HiddenClass, thus violating `JSObject::forEachOwnPropertyWhile`'s
precondition.Therefore, the fix is to use the correct, but slower path,
whenever there is an accessor on the object. This code path is more 1:1
to the spec. At the beginning, it will simply find all the properties
of the source object using `JSObject::getOwnPropertyKeys` and then
iterate on those, rather than use `JSObject::forEachOwnPropertyWhile`.
This way, we are allowed to manipulate the source object however we
wish, which is necessary in order to call `getNamedPropertyValue_RJS`.

Original Reviewed By: neildhar

Original Revision: D41701871

Reviewed By: tmikov

Differential Revision: D44770959

fbshipit-source-id: 8a04ee1b2894972066ed1e8fac36558d2c6dd371