Cache StringPrimitive JenkinsHash
Summary:
Instead of recomputing the StringPrimitive's hash repeatedly,
we store it in `StringPrimitive` after the first calculation.
Because the hash can be 0, we store a bit in the length field to know
whether the hash has been cached.

This can be used in places like Map and IdentifierTable
to avoid unnecessary hash computations.

Reviewed By: neildhar

Differential Revision: D44717626

fbshipit-source-id: 8a3e22418d154c717b71b6485cd281162497c54f