Mitigate RCEs caused by conversion to double
Summary:
This diff is a port of D43363964 for Static Hermes.

Note that unlike the original, it does not change the trustedness
of calls. That is, any existing calls to `encodeNumberValue` and
`encodeDoubleValue` are converted to
`encodeTrustedNumberValue`.

For reference:

Original Author: luigiconiglio@meta.com
Original Git: 83e2cdf0aa4096de6cf9e61964a0ba882b63f41d

# Context

Hermes method to convert raw numeric values to NaN-boxed Hermes values
(`HermesValue::encodeDouble` and `HermesValue::encodeNumber`) are
unsafe by default. These functions allow potential NaN values to be
encoded in an HermesValue and, when provided an untrusted input, can
result in HermesValues encoding objects with arbitrary addresses.
This unsafe behaviour has been the root cause of security incidents in
the past.

# Mitigation

This diffs, converges the current encoding API to a safer and simpler
one.

The old API **(unsafe functions in bold)**:
- `HermesValue::encodeUntrustedDoubleValue`
- **`HermesValue::encodeDoubleValue`**
- `HermesValue::encodeUntrustedNumberValue`
- **`HermesValue::encodeNumberValue` + overloads**

Is now composed of only two functions:
- `HermesValue::encodeUntrustedNumbedrValue`
- **`HermesValue::encodeTrustedNumberValue`**

The new naming makes unsafe function explicit.

Reviewed By: tmikov

Differential Revision: D44812687

fbshipit-source-id: 97c3c187063e6f9f3e896f82c84c5d124b2fbf67