Add native stack checking to Runtime
Summary:
Add a flag to control whether the true native stack checking is enabled.
This controls the VM behavior, defaulted to on but it'll error when
trying to compile for Windows.

Add a function to `Runtime` to check for native stack overflow.
The function is split:
* a fast inline check which does two simple comparisons
* a slow check that resets the stack bounds if the fast check fails

The fast inline check may erroneously return `false`,
but that will only happen if a new thread overlaps the old saved stack.

The amount of gap we're allowed at the end of the stack is configured
via `RuntimeConfig`.

This new mechanism will replace the `nativeCallFrameDepth_` mechanism,
which was in use until now but was conservative by design.
The RAII classes for using it are updated to use the true check,
if it is possible.

Some tests were checking for native stack overflow.
They need to be updated to grow the stack larger and actually cause an
overflow.

Reviewed By: neildhar

Differential Revision: D41601035

fbshipit-source-id: abad8b4a7f1d0868c39018ce4386ddb467a6abc3