Add shermes support for native stack checking
Summary:
Add an option `-check-native-stack`.

`shermes` will output a call to check the native stack for overflow.
For now, this results in an out-of-line call into `Runtime`,
but it can be updated in the future to inline the fast path.
We don't do that yet because SH doesn't have access to Runtime fields.

The option requires that the linked VM was compiled with
`HERMES_CHECK_NATIVE_STACK` enabled.

Reviewed By: tmikov

Differential Revision: D41601036

fbshipit-source-id: 33a5c4ca0b6d884ae257e59b6a3778a7e40ddb22