Infer types of array literals
Summary:
In order to distinguish typed and untyped array literals, IRGen will
use type information added by the `FlowChecker`. This diff adds
inference for the types of array literals by trying the following:

1. First it will look if the array is being assigned to a surrounding
declaration with an array type. If so, it will deduce that as the type
of the array.
2. Otherwise, it will look for a surrounding cast and use that.
3. If the above fail, it will deduce the array's element type to be a
union of its elements.

For now assume any spread element that isn't an array just produces
`any`. The implementation also relies on the implementation of
`SpreadElementNode` to enforce any necessary implicit casts, at
least for now.

Reviewed By: avp

Differential Revision: D44340611

fbshipit-source-id: 745fe5208d5aec3add38f121b27215563b09351a