Supply the input filename as the first parameter to cc
Summary:
GCC is sensitive to the order in which the source file is provided
relative to the libraries we are linking against.

Looking at the linker invocations from the two compilers, it looks like
GCC uses the `--as-needed` flag when linking any user provided
libraries, whereas clang does not. So any shared libraries specified
before the source file end up being dropped by the linker when using
GCC.

Reviewed By: tmikov

Differential Revision: D45171189

fbshipit-source-id: d1b9df8555516edf8fe2b06535ed127287700dee