Replace existing uses of getChangedOperands
Summary:
All read-only accesses of stack locations must go through
`LoadStackInst`, so any other instruction that takes an
`AllocStackInst` must be able to write to it. This means that instead
of checking `getChangedOperands`, we can simply check for operands that
are `AllocStackInst` to find writable stack operands.

This check is slightly more general than `getChangedOperands`, since the
latter is only intended to be used after the IR has been lowered to
bytecode. In practice, this means that `StoreStackInst` will also
trigger this check, although all existing usages occur after it has been
lowered out.

As an optimisation, we can only look at the operands of instructions
that have `WriteStack` set in their side effects, as opposed to looking
at all non-`LoadStackInst` instructions. Note that conditioning on
`WriteStack` automatically excludes `LoadStackInst`.

Reviewed By: tmikov

Differential Revision: D45172591

fbshipit-source-id: ec77a4f3be1fa7a81c435c5fff8c86a6dddb02bc