Remove usage of SimpleCallGraphProvider and delete it
Summary:
Switch TypeInference to using information gathered by
`FunctionAnalysis`, and delete `SimpleCallGraphProvider`.

The primary motivation here is to reduce the number of ways in which
calls and closures can be analysed, to reduce mental overhead when
making changes like the upcoming changes to store `new.target` as an
operand of `BaseCallInst`.

This results in improvements in some cases, because
`FunctionAnalysis` is better at tracking closures.

However it also results in regressions:
1. `SimpleCallGraphProvider` did not handle loose mode functions as
always escaping. This means that it could perform type inference on
their parameters. So the new version is technically more correct in
this regard.
2. `SimpleCallGraphProvider` was able to associate multiple target
functions with a callsite. However, `FunctionAnalysis` only allows for
a single target function.

This also deletes the object property analysis, which seems to be
unused anyway.

Reviewed By: avp

Differential Revision: D45198389

fbshipit-source-id: 3338c41c0ee8c3e8be10c4c0eca3b6fe9e6f34d7