Specialise isDirectCallee for FunctionAnalysis
Summary:
Now that `FunctionAnalysis` is the only remaining user of
`isDirectCallee`, move it there. It is a fairly specialised function to
begin with, and having it in the context it is actually used is
helpful.

More importantly, we can further specialise the function for its
purpose in `FunctionAnalysis`. Since the target IR `Function` is always
known, we can pass that in as an additional parameter. As noted in the
comments, this can be used to tell whether passing the closure as an
argument to the function truly causes it to escape.

This information is used later in the stack to refine how we handle
`new.target` in `FunctionAnalysis`.

Reviewed By: avp

Differential Revision: D45205810

fbshipit-source-id: a5853761354ca1456b073a01d6e31cc92b7bcf4a