Add ThrowTypeErrorInst
Summary:
Add the ability to throw a type error with a specified error message.
This will be used for const enforcement in upcoming diffs.

Note that since const errors are only thrown in test262 mode, there is
no need to optimise for binary size by creating a special instruction
just to throw it.

Reviewed By: avp

Differential Revision: D45557201

fbshipit-source-id: c72d00f0ab237cf9e926f2f020f46e8a30d61be7