Handle recursive types during comparison
Summary:
The `Type::compare` function needs to avoid infinite recursion
when comparing types such as:
```
type A = A[];
type B = B[];
```

It also needs to compare all of these equal to each other:
```
type A = A[];

type B = C[];
type C = B[];
```

To account for this, we add visited sets for `Type *`.
Either the comparison will diverge before we cycle on both types,
or visiting two cycles will mean that the two types must be equal.

Reviewed By: tmikov

Differential Revision: D45189173

fbshipit-source-id: c1a280da38b7fee4d44f922cbaa335f3d4a196c7